+++ tiledesk-client/CHANGELOG.md
@@ -1,11 +1,19 @@
 # Client library for Tiledesk nodejs APIs client
 
+### 0.7.11 - online
+- - bug fix - some projectId to project_id
+
+### 0.7.10 - online
+- removed token as explicit parameter in some function calls. token is now an optional constructor parameter. It's also supported by "options" parameter in some function calls.
+- added customAuthentication()
+
 ### 0.7.9 - online
-- removed project_id as explicit parameter in some function calls. project_id is now an optional constructor parameter. Added "options" parameter some function calls.
+- Added "options" parameter some function calls.
+- removed project_id as explicit parameter in some function calls. project_id is now an optional constructor parameter. It's also supported by "options" parameter in some function calls.
 - added customAuthentication()
 
 ### 0.7.8 - online
-- bug fix- Added missing jwt_token() to sendMessage and more.
+- bug fix - Added missing jwt_token() to sendMessage and more.
 
 ### 0.7.7 - online
 - bug fix: APIURL